// src/app/Shared/Interfaces/cartdetails.ts
export interface Cartdetails {
  status: string
  message: string
  numOfCartItems: number
  data: Data
}

export interface Data {
  _id: string
  cartOwner: string
  products: Product[]
  createdAt: string
  updatedAt: string
  __v: number
  totalCartPrice: number
}

export interface Product {
  count: number
  _id: string
  product: string
  price: number
}